Create getStreamlitWheelFileName()
