+++ packages/desktop/scripts/build_bin.js
@@ -4,7 +4,7 @@ import path from "node:path";
 import fs from "node:fs";
 import { fileURLToPath } from "node:url";
 import esbuild from "esbuild";
-import { getStreamlitVersion } from "@stlite/devutils";
+import { getStreamlitWheelFileName } from "@stlite/devutils";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -13,8 +13,6 @@ const __dirname = path.dirname(__filename);
 
 const infile = path.resolve(__dirname, "../bin-src/dump_artifacts/index.ts");
 const outfile = path.resolve(__dirname, "../bin/dump_artifacts.js");
-
-const streamlitVersion = getStreamlitVersion();
 
 esbuild
   .build({
@@ -32,10 +30,10 @@ esbuild
       "process.env.PATH_FROM_SCRIPT_TO_BUILD": JSON.stringify("../build"),
       "process.env.PATH_FROM_SCRIPT_TO_WHEELS": JSON.stringify("../wheels"),
       "process.env.STREAMLIT_WHEEL_FILE_NAME": JSON.stringify(
-        `streamlit-${streamlitVersion}-cp312-none-any.whl`,
+        getStreamlitWheelFileName(),
       ),
       "process.env.STLITE_LIB_WHEEL_FILE_NAME": JSON.stringify(
-        `stlite_lib-0.1.0-py3-none-any.whl`,
+        "stlite_lib-0.1.0-py3-none-any.whl",
       ),
     },
     outfile,
