+++ packages/devutils/src/index.js
@@ -1,21 +1,6 @@
-import { execSync } from 'node:child_process';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
+import { getStreamlitVersion } from "./get_streamlit_version.js";
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-export function getStreamlitVersion() {
-  try {
-    const version = execSync("uv run python get_streamlit_version.py", {
-      cwd: __dirname,
-      encoding: "utf8",
-    }).trim();
-    return version;
-  } catch (error) {
-    throw new Error(`Failed to get Streamlit version: ${error.message}`);
-  }
-}
-
-if (import.meta.url === `file://${process.argv[1]}`) {
-  console.log(getStreamlitVersion());
+export function getStreamlitWheelFileName() {
+  const streamlitVersion = getStreamlitVersion();
+  return `streamlit-${streamlitVersion}-cp312-none-any.whl`;
 }
