+++ packages/sharing/vite.config.ts
@@ -23,83 +23,79 @@ import wasm from "vite-plugin-wasm";
 import topLevelAwait from "vite-plugin-top-level-await";
 
 import path from "path";
-import { getStreamlitVersion } from "@stlite/devutils";
+import { getStreamlitWheelFileName } from "@stlite/devutils";
 
 const BUILD_AS_FAST_AS_POSSIBLE =
   process.env.BUILD_AS_FAST_AS_POSSIBLE || false;
 
-export default defineConfig(({ mode }) => {
-  const streamlitVersion = getStreamlitVersion();
-
-  return {
-    plugins: [
-      react({
-        // jsxImportSource: "@emotion/react",
-        // plugins: [["@swc/plugin-emotion", {}]],
-        babel: {
-          plugins: ["@emotion/babel-plugin"],
-        },
-      }),
-      viteTsconfigPaths(),
-      wasm(),
-      topLevelAwait(),
-    ],
-    resolve: {
-      alias: {
-        "@streamlit/lib/src": path.resolve(
-          __dirname,
-          "../../streamlit/frontend/lib/src",
-        ),
-        "@streamlit/lib": path.resolve(
-          __dirname,
-          "../../streamlit/frontend/lib/src",
-        ),
-        "stlite_lib.whl": path.resolve(
-          __dirname,
-          "../kernel/py/stlite-lib/dist/stlite_lib-0.1.0-py3-none-any.whl",
-        ),
-        "streamlit.whl": path.resolve(
-          __dirname,
-          `../kernel/py/streamlit/lib/dist/streamlit-${streamlitVersion}-cp312-none-any.whl`,
-        ),
+export default defineConfig(({ mode }) => ({
+  plugins: [
+    react({
+      // jsxImportSource: "@emotion/react",
+      // plugins: [["@swc/plugin-emotion", {}]],
+      babel: {
+        plugins: ["@emotion/babel-plugin"],
       },
-    },
-    assetsInclude: ["**/*.whl"],
-    optimizeDeps: {
-      exclude: ["parquet-wasm"],
-    },
-    worker: {
-      format: "es",
-    },
-    define: {
-      "process.env.NODE_ENV": JSON.stringify(mode),
-      EDITOR_APP_ORIGIN: JSON.stringify(process.env.EDITOR_APP_ORIGIN),
-      EDITOR_APP_ORIGIN_REGEX: JSON.stringify(
-        process.env.EDITOR_APP_ORIGIN_REGEX,
+    }),
+    viteTsconfigPaths(),
+    wasm(),
+    topLevelAwait(),
+  ],
+  resolve: {
+    alias: {
+      "@streamlit/lib/src": path.resolve(
+        __dirname,
+        "../../streamlit/frontend/lib/src",
+      ),
+      "@streamlit/lib": path.resolve(
+        __dirname,
+        "../../streamlit/frontend/lib/src",
+      ),
+      "stlite_lib.whl": path.resolve(
+        __dirname,
+        "../kernel/py/stlite-lib/dist/stlite_lib-0.1.0-py3-none-any.whl",
+      ),
+      "streamlit.whl": path.resolve(
+        __dirname,
+        `../kernel/py/streamlit/lib/dist/${getStreamlitWheelFileName()}`,
       ),
     },
-    server: {
-      open: false,
-      port: 3000,
-      fs: {
-        allow: ["../.."],
-      },
+  },
+  assetsInclude: ["**/*.whl"],
+  optimizeDeps: {
+    exclude: ["parquet-wasm"],
+  },
+  worker: {
+    format: "es",
+  },
+  define: {
+    "process.env.NODE_ENV": JSON.stringify(mode),
+    EDITOR_APP_ORIGIN: JSON.stringify(process.env.EDITOR_APP_ORIGIN),
+    EDITOR_APP_ORIGIN_REGEX: JSON.stringify(
+      process.env.EDITOR_APP_ORIGIN_REGEX,
+    ),
+  },
+  server: {
+    open: false,
+    port: 3000,
+    fs: {
+      allow: ["../.."],
     },
-    build: {
-      outDir: "build",
-      sourcemap: !BUILD_AS_FAST_AS_POSSIBLE,
-      rollupOptions: {
-        output: {
-          // Ensure that *.whl files don't include the hash part to comply with the file name convention
-          // defined in https://packaging.python.org/en/latest/specifications/binary-distribution-format/#file-name-convention.
-          assetFileNames: (assetInfo) => {
-            if (assetInfo.name?.endsWith(".whl")) {
-              return "assets/[name][extname]";
-            }
-            return "assets/[name]-[hash][extname]";
-          },
+  },
+  build: {
+    outDir: "build",
+    sourcemap: !BUILD_AS_FAST_AS_POSSIBLE,
+    rollupOptions: {
+      output: {
+        // Ensure that *.whl files don't include the hash part to comply with the file name convention
+        // defined in https://packaging.python.org/en/latest/specifications/binary-distribution-format/#file-name-convention.
+        assetFileNames: (assetInfo) => {
+          if (assetInfo.name?.endsWith(".whl")) {
+            return "assets/[name][extname]";
+          }
+          return "assets/[name]-[hash][extname]";
         },
       },
     },
-  };
-});
+  },
+}));
